API 추출 (20.11.02)
