+++ TeamProject/static/js/hw/main.js
@@ -1,76 +1,38 @@
-import * as URL from "/static/js/config.js"
+// --------------- api import ----------------
+import * as mainPageApi from '/static/js/hw/api/main_page.js';
 
-//------------------ 베스트 게시글 FetchAPI ------------------
-function get_bestpost_FetchAPI() {
-
-	const get_bestpost_url = URL.BEST_POST;
-	fetch(get_bestpost_url, {
-			method: "GET",
-			headers: {
-				'Accept': 'application/json',
-				'Content-Type': 'application/json',
-			}
-		})
-		.then(res => res.json())
-		.then((res) => {
-			best_post_init(res);
-		})
-}
-get_bestpost_FetchAPI();
-
-// ----------------- 베스트 게시글 -----------------
+mainPageApi.get_best_post();
 function best_post_init(res) {
-	
-	for(let pl of res){
+
+	for (let pl of res) {
 		// div element 생성하고 board 클래스 추가해준다.
 		const post = document.createElement("div");
 		post.classList.add("best_post");
-		
+
 		// post에 들어갈 내용인 in_post이다. 받아온 post_list에서 게시판 이름과 글 제목을 ${}를 통해 삽입해준다.
 		const in_post = `<span class="best_post_board_name">[${pl.board_name}]</span><span class="post_title">${pl.subject}</span><span class="best_post_icons"> <i class="far fa-thumbs-up"></i> ${pl.like_num} <i class="far fa-comment"></i> ${pl.comment_num}</span>`
 		post.innerHTML = in_post;
 
-		post.addEventListener("click",function(){
-			location.href=`post#${pl.board_id}#postinfo#${pl.id}`; //페이지 이동
+		post.addEventListener("click", function () {
+			location.href = `post#${pl.board_id}#postinfo#${pl.id}`; //페이지 이동
 		})
-		
+
 		// best_board div에 삽입해준다.
 		document.querySelector(".best_post_container").appendChild(post);
 	}
 
-	
 	const post_title = document.querySelectorAll(".post_title");
-	
 	// 베스트 게시글의 제목을 모두 불러와서 일정 글자가 넘으면 일정 글자 이후로 ... 으로 바꿔줌
 	for (let pt of post_title) {
 		if (pt.innerText.length > 12) {
 			pt.innerText = pt.innerText.substr(0, 12) + '...';
 		}
 	}
-
 }
 
-//------------------ 베스트 게시판 FetchAPI ------------------
-function get_bestboard_FetchAPI() {
-
-	const get_bestboard_url = URL.BEST_BOARD;
-	fetch(get_bestboard_url, {
-			method: "GET",
-			headers: {
-				'Accept': 'application/json',
-				'Content-Type': 'application/json',
-			}
-		})
-		.then(res => res.json())
-		.then((res) => {
-			best_board_init(res);
-		})
-}
-get_bestboard_FetchAPI();
-
-// ------------------ 베스트 게시판 ----------------------
+mainPageApi.get_best_board();
 function best_board_init(res) {
-	
+
 	const slider = document.querySelector(".slider");
 
 	for (let bb of res) {
@@ -79,25 +41,25 @@ function best_board_init(res) {
 		slide.classList.add("slide");
 
 		// 이미지가 없다면 디폴트 이미지 넣어줌
-		if(bb.board_image == null){
+		if (bb.board_image == null) {
 			slide.innerHTML = `<img src="../static/img/main_img/board_default.png" class="s_img">
 		<div>${bb.board_name}</div>`;
 		}
 		// 이미지 있으면 그대로 게시판 이미지, 이름 넣어줌
-		else{
+		else {
 			slide.innerHTML = `<img src="../static/img/board_img/${bb.board_image}" class="s_img">
-			<div>${bb.board_name}</div>`;	
+			<div>${bb.board_name}</div>`;
 		}
 
 		// 마우스 hovering
-		slide.addEventListener("mouseenter",function(){
+		slide.addEventListener("mouseenter", function () {
 			slide.style.opacity = "1";
 		})
-		slide.addEventListener("mouseleave",function(){
+		slide.addEventListener("mouseleave", function () {
 			slide.style.opacity = "0.7";
 		})
-		slide.addEventListener("click",()=>{
-			location.href=`post#${bb.id}#postmain`
+		slide.addEventListener("click", () => {
+			location.href = `post#${bb.id}#postmain`
 		})
 		slider.appendChild(slide);
 	}
@@ -113,7 +75,6 @@ function best_board_init(res) {
 
 	// 좌우 버튼 누르면 긴 slider 자체가 좌우로 이동하는 방식
 	left_btn.addEventListener("click", function () {
-
 		index--;
 		if (index < 0) {
 			index = slides.length - 6;
@@ -125,11 +86,9 @@ function best_board_init(res) {
 
 		const bar_move = index * (100 / 5);
 		bar.style.left = bar_move + '%';
-
 	})
 
 	right_btn.addEventListener("click", function () {
-
 		index++;
 		if (index > slides.length - 6) {
 			index = 0;
@@ -141,70 +100,48 @@ function best_board_init(res) {
 
 		const bar_move = index * (100 / 5);
 		bar.style.left = bar_move + '%';
-
 	})
 
 }
 
-
 // --------------------------------------------- 카테고리 ------------------------------------------ //
-//------------- 카테고리 반환 FetchAPI --------------
-function get_category_FetchAPI() {
-
-	const get_category_url = URL.CATEGORY_INFO;
-	fetch(get_category_url, {
-			method: "GET",
-			headers: {
-				'Accept': 'application/json',
-				'Content-Type': 'application/json',
-			}
-		})
-		.then(res => res.json())
-		.then((res) => {
-			category_init(res);
-		})
-}
-get_category_FetchAPI();
-
-// ------- 카테고리 init --------
+mainPageApi.get_all_category();
 function category_init(res) {
-
 	// --------------- 카테고리 컨테이너 생성 -------------
-	let first=0;
-	for(let cg of res){
+	let first = 0;
+	for (let cg of res) {
 		const category = document.createElement("div");
 		category.classList.add("category");
 		// 첫번째 카테고리는 바로 보여야되기 때문에 active 클래스를 넣어준다.
-		if(first==0){
+		if (first == 0) {
 			category.classList.add("active");
 			first++;
 		}
-	
+
 		category.innerHTML = `<div class="category_name" category_id="${cg.id}">${cg.category_name}</div>
 		<div class="board_container"></div>
 		<div class="board_page_container" id="pagination"></div><div class="b_btn_left b_btn">
 		<img src="../static/img/main_img/arrow-left.png" alt="" class="b_btn_img">
-	</div>
-	<div class="b_btn_right b_btn">
-		<img src="../static/img/main_img/arrow-right.png" alt="" class="b_btn_img">
-	</div>`
+		</div>
+		<div class="b_btn_right b_btn">
+			<img src="../static/img/main_img/arrow-right.png" alt="" class="b_btn_img">
+		</div>`
 
 		document.querySelector(".category_container").appendChild(category);
 	}
-	
+
 	// 메인페이지에 들어오면 하단에 보여질 첫번째 카테고리의 게시판들을 먼저 init 해준다.
 	const first_category_id = document.querySelector(".active").childNodes[0].getAttribute("category_id");
-	get_board_FetchAPI(first_category_id);
+	mainPageApi.get_all_board_in_category(first_category_id);
 
 	// ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ 카테고리 컨테이너 생성 완료
-	
 
 	// ---------- 왼쪽, 오른쪽 버튼 이벤트리스너 init 부분 ----------
 	const category_left_btns = document.querySelectorAll(".b_btn_left");
 	const category_right_btns = document.querySelectorAll(".b_btn_right");
 	const categorys = document.querySelectorAll(".category");
 
-	for(let category_left_btn of category_left_btns){
+	for (let category_left_btn of category_left_btns) {
 		category_left_btn.addEventListener("click", function () {
 
 			// 현재 active된 카테고리를 불러온다.
@@ -217,13 +154,13 @@ function category_init(res) {
 					categorys[categorys.length - 1].classList.add("active");
 				}
 			}
-	
+
 			// 바뀐 카테고리의 아이디를 가져와서 해당 카테고리의 게시판들을 받아 init 해준다.
 			const changed_category_id = document.querySelector(".active").childNodes[0].getAttribute("category_id");
-			get_board_FetchAPI(changed_category_id);
+			mainPageApi.get_all_board_in_category(changed_category_id);
 		})
 	}
-	
+
 	for (let category_right_btn of category_right_btns) {
 		category_right_btn.addEventListener("click", function () {
 
@@ -240,31 +177,10 @@ function category_init(res) {
 
 			// 바뀐 카테고리의 아이디를 가져와서 해당 카테고리의 게시판들을 받아 init 해준다.
 			const changed_category_id = document.querySelector(".active").childNodes[0].getAttribute("category_id");
-			get_board_FetchAPI(changed_category_id);
+			mainPageApi.get_all_board_in_category(changed_category_id);
 		})
 	}
 }
-
-
-// ------------------------------------- 게시판 ---------------------------------------------
-
-// ------------- 해당 카테고리에 속한 게시판 반환 FetchAPI --------------
-function get_board_FetchAPI(category_id) {
-
-	const get_board_url = URL.BOARDS_IN_CATEGORY + category_id;
-	fetch(get_board_url, {
-			method: "GET",
-			headers: {
-				'Accept': 'application/json',
-				'Content-Type': 'application/json',
-			}
-		})
-		.then(res => res.json())
-		.then((res) => {
-			board_in_category_pagination(res);
-		})
-}
-
 // ------------- 한 카테고리에 들어가는 모든 게시판들을 paging 하여 보여주는 함수 ----------------
 function board_in_category_pagination(board_list) {
 
@@ -290,16 +206,15 @@ function board_in_category_pagination(board_list) {
 			let item_element = document.createElement('span');
 			item_element.classList.add('board');
 
-			if(item.board_image == "" || item.board_image == null){
+			if (item.board_image == "" || item.board_image == null) {
 				item_element.innerHTML = `<img src="../static/img/main_img/board_default.png" class="category_board_image"> ${item.board_name}`;
-			}
-			else{
+			} else {
 				item_element.innerHTML = `<img src="../static/img/board_img/${item.board_image}" class="category_board_image"> ${item.board_name}`;
 			}
 
 			// 해당 게시판을 누를 시 링크 이동 리스너
-			item_element.addEventListener("click",function(){
-				location.href=`post#${item.id}#postmain`;
+			item_element.addEventListener("click", function () {
+				location.href = `post#${item.id}#postmain`;
 			})
 
 			container.appendChild(item_element);
@@ -314,7 +229,7 @@ function board_in_category_pagination(board_list) {
 		for (let i = 1; i < page_count + 1; i++) {
 			let btn = PaginationButton(i, board_list);
 			container.appendChild(btn);
-		}	
+		}
 	}
 
 	function PaginationButton(i, board_list) {
@@ -340,4 +255,11 @@ function board_in_category_pagination(board_list) {
 	DisplayList(board_list, board_container, show_cnt, current_page);
 	SetupPagination(board_list, page_container, show_cnt);
 
+}
+
+export {
+	best_post_init,
+	best_board_init,
+	category_init,
+	board_in_category_pagination
 }
