+++ TeamProject/templates/main.html
@@ -89,15 +89,7 @@
     </div>
 
     <div class="b_container aos-animate aos-init" data-aos="fade-left" data-aos-duration="1000">
-        <!-- <div class="b_btn_left b_btn">
-            <img src="../static/img/main_img/arrow-left.png" alt="" class="b_btn_img">
-        </div>
-        <div class="b_btn_right b_btn">
-            <img src="../static/img/main_img/arrow-right.png" alt="" class="b_btn_img">
-        </div> -->
-        <div class="category_container">
-
-        </div>
+        <div class="category_container"></div>
     </div>
 
     <script src="https://unpkg.com/aos@2.3.1/dist/aos.js"></script>
@@ -105,7 +97,8 @@
         AOS.init();
     </script>
     <script type="module" src="/static/js/config.js"></script>
-    <script type="module" src="/static/js/main.js" ></script>
+    <script type="module" src="/static/js/hw/api/main_page.js"></script>
+    <script type="module" src="/static/js/hw/main.js" ></script>
     <script type="module" src="/static/js/auth.js" ></script>
 
 </body>
