+++ TeamProject/templates/manager.html
@@ -130,7 +130,10 @@
     <script type="module" src="/static/js/hw/components/modal/add_user_blacklist.js"></script>
     <script type="module" src="/static/js/hw/components/modal/modify_board.js"></script>
     <script type="module" src="/static/js/hw/components/modal/modify_user.js"></script>
-    <script type="module" src="/static/js/manager.js"></script>
+    <script type="module" src="/static/js/hw/api/management/board_and_category.js"></script>
+    <script type="module" src="/static/js/hw/api/management/report.js"></script>
+    <script type="module" src="/static/js/hw/api/management/user.js"></script>
+    <script type="module" src="/static/js/hw/manager.js"></script>
 
 </body>
