+++ TeamProject/templates/mypage.html
@@ -43,7 +43,7 @@
 
     <script type="module" src="/static/js/config.js"></script>
     <script type="module" src="/static/js/auth.js" ></script>
-    <script type="module" src="/static/js/mypage.js" ></script>
+    <script type="module" src="/static/js/hw/mypage.js" ></script>
 
 </body>
